% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiemposCMTC.R
\name{tiemposOcupacionCMTC}
\alias{tiemposOcupacionCMTC}
\title{Cargue la matriz generadora, el t a evaluar (debe coincidir con las tasas de la matriz Q) y el alpha o vector de probabilidad inicial.}
\usage{
tiemposOcupacionCMTC(matrizQ, t, alpha)
}
\arguments{
\item{matrizQ}{Matriz generadora de la CMTC}

\item{t}{tiempo a evaluar (debe estar en las mismas unidades que están las tasas de la matriz generadora)}

\item{alpha}{Vector de probabilidad inicial}
}
\value{
El número de transiciones aproximadas que transcurren en el intervalo de tiempo dado por parámetro y la matriz de ocupación (M^t)
}
\description{
Esta función permite calcular la aproximación de la matriz de ocupación dada una matriz generadora (matrizQ), un tiempo a evaluar (t) y el vector de probabilidad inicial (alpha)
}
